Rename file and variable
